rag/requirements.txt
python-dotenv
openai
azure-cosmos
pandas
numpy
sklearn